# Reduce allocations for multipling LazyTensor of sparse and dense