--- a/src/operators_lazytensor.jl
+++ b/src/operators_lazytensor.jl
@@ -648,7 +648,7 @@ function _gemm_puresparse(alpha, op::Matrix, h::LazyTensor{B1,B2,F,I,T}, beta, r
         rmul!(result, beta)
     end
     N_k = length(h.basis_r.bases)
-    shape, strides_j, strides_k = _get_shape_and_srtides(h)
+    shape, strides_j, strides_k = _get_shape_and_strides(h)
     _gemm_recursive_dense_lazy(1, N_k, 1, 1, alpha*h.factor, shape, strides_k, strides_j, h.indices, h, op, result)
 end
 
@@ -659,11 +659,11 @@ function _gemm_puresparse(alpha, h::LazyTensor{B1,B2,F,I,T}, op::Matrix, beta, r
         rmul!(result, beta)
     end
     N_k = length(h.basis_l.bases)
-    shape, strides_j, strides_k = _get_shape_and_srtides(h)
+    shape, strides_j, strides_k = _get_shape_and_strides(h)
     _gemm_recursive_lazy_dense(1, N_k, 1, 1, alpha*h.factor, shape, strides_k, strides_j, h.indices, h, op, result)
 end
 
-function _get_shape_and_srtides(h)
+function _get_shape_and_strides(h)
     shape_l, shape_r = _comp_size(h.basis_l), _comp_size(h.basis_r)
     shape = min.(shape_l, shape_r)
     strides_j, strides_k = _strides(shape_l), _strides(shape_r)