Following suggestion from @amilsted regarding https://github.com/qojulia/QuantumOptics.jl/issues/352

`shape`, `strides_j` and `strides_k` in `_gemm_puresparse` function are implemented as `Tuple` to reduce allocations.