@Krastanov 2023-02-24
This is a great improvement, thank you for taking initiative on it! Could you share a `BenchmarkTools.@benchmark` comparison before/after your change?