@Krastanov 2023-02-24
Here is a quick before/after.

Before:

```
julia> b = FockBasis(20)
       op1 = destroy(b)
       op2 = create(b)
       op3 = op1⊗op1
       lop = LazySum(LazyTensor(basis(op3),[1,2],[op1,op2]),op3)
       s = basisstate(b,5)
       s = s⊗s
       ss = copy(s)
       @benchmark QuantumOptics.mul!(s,lop,s)
BenchmarkTools.Trial: 10000 samples with 7 evaluations.
 Range (min … max):  4.354 μs …   7.528 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     4.405 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   4.441 μs ± 117.392 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%
 Memory estimate: 432 bytes, allocs estimate: 7.

```

After:

```
julia> b = FockBasis(20)
       op1 = destroy(b)
       op2 = create(b)
       op3 = op1⊗op1
       lop = LazySum(LazyTensor(basis(op3),[1,2],[op1,op2]),op3)
       s = basisstate(b,5)
       s = s⊗s
       ss = copy(s)
       @benchmark QuantumOptics.mul!(s,lop,s)
BenchmarkTools.Trial: 10000 samples with 7 evaluations.
 Range (min … max):  4.344 μs …  7.348 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     4.384 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   4.400 μs ± 92.505 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%
 Memory estimate: 192 bytes, allocs estimate: 4.
```

The change in performance is quite minimal in this microbenchmark (no runtime duration change, but half the allocations), but still better than status quo.