@AmitRotem 2023-02-24
```
# basis
sub_system = FockBasis(3)
number_of_sub_systems = 5
full_system = sub_system^number_of_sub_systems
# state
ψ0 = randstate(full_system);
# hamiltonian
H_sum_tensor_sparse = let ops1, ops2, n=length(sub_system), r=0.5
    ops1 = (Operator(sub_system, sprandn(ComplexF64,n,n,r)) for k=1:number_of_sub_systems)
    ops1 = ops1.+dagger.(ops1)
    ops2 = (Operator(sub_system, sprandn(ComplexF64,n,n,r)) for k=1:number_of_sub_systems-1, j=1:2)
    ops2 = ops2.+dagger.(ops2)
    LazySum((LazyTensor(full_system, j, op) for (j,op)=enumerate(ops1))...)+
     LazySum((LazyTensor(full_system, (j,j+1), tuple(op...)) for (j,op)=enumerate(eachrow(ops2)))...)
end

Ht(H) = (_,_)->H # just be because schroedinger won't propagate operators

# propagate state
@time timeevolution.schroedinger_dynamic((0.0, 1.0), ψ0state, Ht(H_sum_tensor_sparse))
@time timeevolution.schroedinger_dynamic((0.0,10.0), ψ0state, Ht(H_sum_tensor_sparse))
# propagate operator
@time timeevolution.schroedinger_dynamic((0.0, 1.0), ψ0op, Ht(H_sum_tensor_sparse))
@time timeevolution.schroedinger_dynamic((0.0,10.0), ψ0op, Ht(H_sum_tensor_sparse))
;
```

Which for the current release gives gives;
```
0.239708 seconds (18.42 k allocations: 1.503 MiB)
1.644933 seconds (168.87 k allocations: 11.341 MiB)
0.356556 seconds (8.47 k allocations: 2.256 MiB)
3.014540 seconds (74.08 k allocations: 8.263 MiB)
```
And for this branch gives gives;
```
0.194773 seconds (11.43 k allocations: 842.953 KiB)
1.789770 seconds (108.41 k allocations: 5.263 MiB)
0.359107 seconds (123 allocations: 1.492 MiB)
3.327338 seconds (123 allocations: 1.492 MiB)
```