@amilsted 2023-02-24
Curious that the operator case allocations go down way more than the state case. Do you know why?