@AmitRotem 2023-02-24
benchmark notebooks
[release](https://github.com/AmitRotem/QuantumOpticsBase.jl/blob/bench/benchmark/release/lazy%20sparse%20release.ipynb)
[this branch](https://github.com/AmitRotem/QuantumOpticsBase.jl/blob/bench/benchmark/main/lazy%20sparse%20main.ipynb)