@AmitRotem 2023-02-24
The only difference between Ket and Operator is the reshape of Ket.data to a Matrix