@AmitRotem 2023-02-24
Also, in the notebooks you can see that propagating operator with a dense hamiltonian has lots of allocations, whereas propagating a state does not. Is that a `LinearAlgebra` issue?