@amilsted 2023-02-24
> The only difference between Ket and Operator is the reshape of Ket.data to a Matrix

Maybe it's a `ReshapedArray()` allocation...

> Also, in the notebooks you can see that propagating operator with a dense hamiltonian has lots of allocations, whereas propagating a state does not. Is that a `LinearAlgebra` issue?

This is https://github.com/JuliaLang/julia/issues/46865, which I have been [trying](https://github.com/JuliaLang/julia/pull/47206) to get fixed!