@AmitRotem 2023-02-25
This reduces the allocation for `Ket`. Not sure if it will be this simple for `Bra`.

But this has a bad effect on LazySum of LazyTensor of dense.