@AmitRotem 2023-03-22
Added dim check related to https://github.com/qojulia/QuantumOpticsBase.jl/issues/74

Also, fix `size` of `AbstractOperator`